AppPackage: allow different owner UIDs/GIDs
By default, when tar is started as root, it tries to keep the same
ownership ids as in the archive. It that fails, then tar exits with
an error.

On the HP Touchpad, the extraction is done in /media/cryptofs, which
is a vfat filesystem, and that doesn't handle UIDs/GIDs.
But more generally, the installer shouldn't trust the package to have
always sane ownership values, and should impose its own values.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>
Upstream-Status: Pending